Having the meta-eval config stuff be loaded whenver the package object is used makes this difficult to use as a library.  Factoring out into a Config object.
