+++ src/main/scala/org/allenai/scholar/metrics/metadata/Main.scala
@@ -6,6 +6,7 @@ import java.io.File
 import java.nio.file.{ Files, Paths }
 
 object Main extends App {
+  import Config._
   /** Run only Grobid's processHeader for now, not fullText.
     * https://github.com/kermitt2/grobid/wiki/Grobid-batch-quick-start
     */
